feat: hs

+++ internal/app/hospitalization_statistics/crawl.go
@@ -13,7 +13,7 @@ import (
 )
 
 type Item struct {
-	Cost      int64  `json:"cost"`
+	Cost      int    `json:"cost"`
 	DeptName  string `json:"deptName"`
 	ItemName  string `json:"itemName"`
 	ItemPrice string `json:"itemPrice"`
